Replace searchtype by matchertype
Determining which search to perform (either base or max_count) is now
based on a field inside the Matcher struct (matcher_type). This is more
convenient as once the Matcher struct has been built, it is already
known what kind of matcher it is. The only thing left to do for
search_matches is to extract that field and see which API to call.