Refactor searcher to hold multiple result types
This commit lays down the foundation for the new -c (--count) flag that
will only show the number of matching lines. This functionality isn't
implemented yet although the new flag has been added to the cli.

The main idea behind this refactoring was to have a more 'general'
result type that can either hold the matches from a regular search
or only the number of matching lines (if the -c flag is specified).
A simple solution here could have been to just add a new field
to the already exsiting SearchResult struct to account for the
matching number of lines. However this approach didn't feel very
satisfying as the regular search and the new 'count' search are two
distinct search types. Therefore this commit introduces an enum
for holding either a SearchResult or CountResult. In both cases,
the results are 'upcasted', via trait Upcast, to a 'general'
result (GenResult). This approach allows, among other things, the new
count methods to be implemented in a very similar fashion as the
search counterparts. Furthermore, the new CounterResult implements a
method for counting matching lines and tests have been adjusted where
necessary.