Implement -c (--count) flag
This commit extends the preparations carried out in a96645a and
finalizes the implementation for the new -c flag. The new flag is
compatible with all other flags. Currently, the only feature missing
is that it's not possible to use this flag without specifying
a PATTERN. For example, counting the number of lines in a file can be
expressed as `grrs -c "" file.txt`. However, in the future it would be
convenient to simplify this to just `grss -c file.txt`

A few additional things have been achieved:
- TODO's in comments
- remove redundant if statements from gen_check mod
- refactor and extend writer mod for printing count results
- improve README.md